fix: move user logic from general to api hook
